Refactor plus more tests
