+++ org.hl7.fhir.publisher.core/src/main/java/org/hl7/fhir/igtools/publisher/Publisher.java
@@ -8576,18 +8576,7 @@ public class Publisher implements IWorkerContext.ILoggingService, IReferenceReso
     return businessVersion == null ? publishedIg.getVersion() : businessVersion;
   }
 
-  protected static String execAndReturnString(String[] cmd, String[] env, File directory) throws IOException, InterruptedException {
-    String output = "";
-    Process p = Runtime.getRuntime().exec(cmd, null, directory);
-    InputStreamReader isr = new InputStreamReader(p.getInputStream());
-    p.waitFor();
-    BufferedReader br = new BufferedReader(isr);
-    String line;
-    while ((line = br.readLine()) != null) {
-      output += line;
-    }
-    return output;
-  }
+
 
   private String getGitStatus() throws IOException {
     File gitDir = new File(Utilities.getDirectoryForFile(configFile));
