+++ org.hl7.fhir.publisher.core/src/test/java/org/hl7/fhir/igtools/publisher/GitUtilitiesTests.java
@@ -5,16 +5,23 @@ import org.junit.jupiter.api.Test;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
+import java.nio.file.Path;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class GitUtilitiesTests {
 	@Test
 	public void testGetGitStatus() {
-		String output = GitUtilities.getGitStatus(new File(System.getProperty("user.dir")));
-		assertTrue(output.length() > 0);
+		String output = GitUtilities.getGitStatus(Path.of(new File(System.getProperty("user.dir")).getAbsolutePath(), "src/test/resources/git/regular-branch/").toFile());
+		assertEquals("HEAD", output.trim());
 	}
 
+	@Test
+	public void testGetGitWorktreeStatus() {
+		String output = GitUtilities.getGitStatus(Path.of(new File(System.getProperty("user.dir")).getAbsolutePath(), "src/test/resources/git/worktree-branch/").toFile());
+		assertEquals("HEAD", output.trim());
+	}
 	@Test
 	public void testGitStatusWhenNotGitDirectory() throws IOException {
 		String output = GitUtilities.getGitStatus(Files.createTempDirectory("emptyNonGitDirectory").toFile());
